pyproject.toml
[tool.poetry]
name = "gpt-assistant"
version = "0.1.0"
description = ""
authors = ["Chris Doucette <chrisdoucette15@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt_assistant"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.4"
pyttsx3 = "^2.90"
speechrecognition = "^3.10.0"
pyaudio = "^0.2.13"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
assistant = "gpt_assistant:main"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
